#58 added file for GATTBarometricPressureTrend
